debian: find firmware: quote regex metacharacters in directory name
The directory name contains the version which may contain characters
with semantics as regex metacharacters, e.g. `+` when building a test
kernel. If they are not quoted, the script does not remove the
directory name from its output, which will cause a build failure in
`fwcheck` even though the firmware did not change.

To avoid that, quote metacharacters in the directory name.

Signed-off-by: Friedrich Weber <f.weber@proxmox.com>
Link: https://lore.proxmox.com/20250428163412.3670859-1-f.weber@proxmox.com